Create automated whatsapp texter
Here i worked with pywhatkit module which is one among the most widely used python libraries. I used one of its attributes after importing it which is 'sendwhatmsg' which automatically sends the text at the mentioned time. So i basically gave 3 parameters which are the contact number i wanna text , the text and time respectively.